+++ individual_test.go
@@ -5,7 +5,7 @@ import "reflect"
 
 func TestNewIndividual(t *testing.T) {
     individual := NewIndividual(3, NewGeneTestFunction)
-    if ! reflect.DeepEqual(individual.Genome, []int{9, 9, 9}) {
+    if ! reflect.DeepEqual(individual.Genome, []float64{9, 9, 9}) {
         t.Fail()
     }
 }
@@ -22,7 +22,7 @@ func TestIndividualCopy(t *testing.T) {
     individual := NewIndividual(3, NewGeneTestFunction)
     individual2 := individual.Copy()
     individual.Genome[2] = 8
-    if ! reflect.DeepEqual(individual2.Genome, []int{9, 9, 9}) {
+    if ! reflect.DeepEqual(individual2.Genome, []float64{9, 9, 9}) {
         t.Fail()
     }
 }
