+++ test_helpers.go
@@ -1,9 +1,9 @@
 package exodus
 
-func NewGeneTestFunction() int {
+func NewGeneTestFunction() float64 {
     return 9
 }
 
-func FitnessTestFunction(genome []int) float64 {
+func FitnessTestFunction(genome []float64) float64 {
     return 1.25
 }
